ci: ensure signing bundles are uploaded alongside
GoReleaser requires that the signature file is created by a signing process, and uploads it alongside the target binary. If the signature file is not unique then it can be overwritten by the signatures of other releases. This causes (recoverable) errors when the release is process.

If the name of the signature file is not aligned with the executable archive, it will also be difficult for the user to determine which signature file to download for a given release.

Unfortunately, this means that the bundle file now incorporated with the archive includes the OS and Arch markers, where the binary does not. This complicates the verification instructions.