fix: include signature bundles only in release archives
This works by naming the bundle file using the
binary name with no OS or architecture
information. This works around the issue with
GoReleaser not working when artifact attributes
are used:
https://github.com/goreleaser/goreleaser/issues/5147

The bundle is part of the release archives, so
builds are verifiable, but they're not published
at the top level of the release. I can live with
this.